+++ tests/index.test.js
@@ -22,6 +22,7 @@ const lc39 = require('@mia-platform/lc39')
 const fs = require('fs')
 const { promisify } = require('util')
 
+const USER_PROPERTIES_HEADER_KEY = 'miauserproperties'
 const USERID_HEADER_KEY = 'userid-header-key'
 const GROUPS_HEADER_KEY = 'groups-header-key'
 const CLIENTTYPE_HEADER_KEY = 'clienttype-header-key'
@@ -30,6 +31,7 @@ const MICROSERVICE_GATEWAY_SERVICE_NAME = 'microservice-gateway'
 
 const baseEnv = {
   USERID_HEADER_KEY,
+  USER_PROPERTIES_HEADER_KEY,
   GROUPS_HEADER_KEY,
   CLIENTTYPE_HEADER_KEY,
   BACKOFFICE_HEADER_KEY,
@@ -68,6 +70,7 @@ tap.test('Plain Custom Service', test => {
         [CLIENTTYPE_HEADER_KEY]: 'CMS',
         [GROUPS_HEADER_KEY]: 'group-name,group-to-greet',
         [USERID_HEADER_KEY]: 'Mark',
+        [USER_PROPERTIES_HEADER_KEY]: JSON.stringify({ prop1: 'value1', prop2: 'value2' }),
       },
     })
 
@@ -77,7 +80,37 @@ tap.test('Plain Custom Service', test => {
     assert.strictSame(JSON.parse(response.payload), {
       userId: 'Mark',
       userGroups: ['group-name', 'group-to-greet'],
+      userProperties: { prop1: 'value1', prop2: 'value2' },
       clientType: 'CMS',
+      backoffice: false,
+    })
+    assert.end()
+  })
+
+  test.test('Access platform "miauserproperties" - when USER_PROPERTIES_HEADER_KEY NOT defined - uses default header key', async assert => {
+    const envarWithoutUserProperties = {
+      ...baseEnv,
+    }
+    delete envarWithoutUserProperties.USER_PROPERTIES_HEADER_KEY
+
+    const fastify = await setupFastify(envarWithoutUserProperties)
+
+    const response = await fastify.inject({
+      method: 'GET',
+      url: '/platform-values',
+      headers: {
+        'miauserproperties': JSON.stringify({ prop1: 'value1', prop2: 'value2' }),
+      },
+    })
+
+    assert.strictSame(response.statusCode, 200)
+    assert.ok(/application\/json/.test(response.headers['content-type']))
+    assert.ok(/charset=utf-8/.test(response.headers['content-type']))
+    assert.strictSame(JSON.parse(response.payload), {
+      userId: null,
+      userGroups: [],
+      userProperties: { prop1: 'value1', prop2: 'value2' },
+      clientType: null,
       backoffice: false,
     })
     assert.end()
@@ -96,6 +129,7 @@ tap.test('Plain Custom Service', test => {
     assert.ok(/charset=utf-8/.test(response.headers['content-type']))
     assert.strictSame(JSON.parse(response.payload), {
       userId: null,
+      userProperties: null,
       userGroups: [],
       clientType: null,
       backoffice: false,