+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added `getUserProperties` to decorated `fastify.Request`, which returns the user properties
+
 ### Changed
 
 - Update @mia-platform/lc39 2.2.2 -> 2.4.0